+++ Demo/CrossDK-demo UIKit/CrossDK.xcframework/ios-arm64_x86_64-maccatalyst/CrossDK.framework/Versions/A/Resources/CHANGELOG.md
@@ -1,7 +1,45 @@
 # CrossDK - CHANGELOG
 
 
+# Q122
+
+
+## Sprint 3
+
+### Feature
+
+* ARP-2665 : Deploy V3.0 SDK containing interstitial overlay
+
+
+## Sprint 2
+
+### Feature
+
+* ARP-2621 : Creation of the interstitial format in UIKit in CrossDK project 
+
+
+## Sprint 1
+
+### Feature
+
+* ARP-2586 : Mock SDK interstitial format 
+
+
 # Q421
+
+
+## Sprint 6
+
+### Feature
+
+* ARP-2555 : Dismiss all overlay when a new one is called
+
+
+## Sprint 5
+
+### Feature
+
+* ARP-2510 : Add an ad type parameter in the SDK integration
 
 
 ## Sprint 4
@@ -11,6 +49,7 @@
 * ARP-2321 : Mock SDK video format
 * ARP-2440 : UIKit Overlay : Objective-C compatibility
 * ARP-2453 : Deploy V2.0 SDK containing video overlay
+
 
 ## Sprint 3
 
@@ -37,6 +76,7 @@
 
 
 # Q321
+
 
 ## Sprint 6